+++ foliant/meta_commands/base.py
@@ -8,13 +8,15 @@ class BaseMetaCommand:
     config_section = ''
     defaults = {}
 
-    def __init__(self, context: dict, logger: Logger):
+    def __init__(self, context: dict, logger: Logger, quiet: bool, debug: bool):
         self.context = context
         self.logger = logger
         self.config = context['config']
         self.project_path = context['project_path']
         options = self.config.get(self.config_section, {}) if self.config_section else {}
         self.options = {**self.defaults, **options}
+        self.quiet = quiet
+        self.debug = debug
 
     def run(self, *args, **kwargs):
         raise NotImplementedError()