+++ foliant/meta_commands/generate/command.py
@@ -3,6 +3,7 @@
 import yaml
 
 from foliant.meta_commands.base import BaseMetaCommand
+from foliant.utils import spinner
 
 from foliant.meta.generate import load_meta
 
@@ -25,14 +26,25 @@ class MetaCommand(BaseMetaCommand):
 
     def run(self):
         self.logger.debug('Meta command generate started')
+        filename = self.options['filename']
+        result = None
+        with spinner(f'Generating metadata', self.logger, self.quiet, self.debug):
+            self._gen_meta()
+            with open(filename, 'w', encoding='utf8') as f:
+                yaml.dump(self.meta.dump(),
+                          f,
+                          default_flow_style=False,
+                          allow_unicode=True,
+                          sort_keys=False)
+            result = filename
 
-        self._gen_meta()
-        with open(self.options['filename'], 'w', encoding='utf8') as f:
-            yaml.dump(self.meta.dump(),
-                      f,
-                      default_flow_style=False,
-                      allow_unicode=True,
-                      sort_keys=False)
-        self.meta.filename = self.options['filename']
+        if result:
+            self.logger.info(f'Result: {result}')
+
+            if not self.quiet:
+                print('─' * 20)
+                print(f'Result: {result}')
+            else:
+                print(result)
 
         self.logger.debug('Meta command generate finished')